+++ pkg/reduce/pnf/processandforward_test.go
@@ -43,7 +43,7 @@ import (
 	"github.com/golang/mock/gomock"
 	functionpb "github.com/numaproj/numaflow-go/pkg/apis/proto/function/v1"
 	"github.com/numaproj/numaflow-go/pkg/apis/proto/function/v1/funcmock"
-	"github.com/numaproj/numaflow/pkg/func-client/clienttest"
+	"github.com/numaproj/numaflow/pkg/sdkclient/clienttest"
 	"github.com/stretchr/testify/assert"
 
 	"github.com/numaproj/numaflow/pkg/isb/testutils"
