+++ pkg/sink-client/client/interface.go
@@ -2,19 +2,9 @@ package client
 
 import (
 	"context"
-	"time"
-
 	sinkpb "github.com/numaproj/numaflow-go/pkg/apis/proto/sink/v1"
 	"google.golang.org/protobuf/types/known/emptypb"
 )
-
-type Datum interface {
-	Keys() []string
-	Value() []byte
-	EventTime() time.Time
-	Watermark() time.Time
-	ID() string
-}
 
 // Client contains methods to call a gRPC client.
 type Client interface {
