+++ pkg/udf/function/uds_grpc_test.go
@@ -27,7 +27,7 @@ import (
 
 	functionpb "github.com/numaproj/numaflow-go/pkg/apis/proto/function/v1"
 	"github.com/numaproj/numaflow-go/pkg/apis/proto/function/v1/funcmock"
-	"github.com/numaproj/numaflow/pkg/func-client/clienttest"
+	"github.com/numaproj/numaflow/pkg/sdkclient/clienttest"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
